requirements.txt
rich==13.7.0
requests==2.31.0
python-dotenv==1.0.1
